Move logic from game_backend Users to champions Campaigns
